placed questions array in separate js file
