+++ src/utils/writeToFile.js
@@ -5,7 +5,6 @@ const writeToFile = (fileName, markdown) => {
   try {
     // write data to new file
     fs.writeFileSync(`./${fileName}.md`, markdown);
-    console.log("Writing to file successful!");
   } catch (error) {
     console.log(error.message);
   }
